Introduce directories and files structure
