+++ modules/clinic/clinic.hpp
@@ -1,0 +1,17 @@
+#pragma once
+
+#include <memory>
+#include <vector>
+#include "staff/istaff.hpp"
+
+class Clinic
+{
+    std::unique_ptr<iStaff> staff_;
+    std::vector<std::unique_ptr<iRoom>> rooms_;
+
+  public:
+    Clinic(std::unique_ptr<iStaff> staff, std::vector<std::unique_ptr<iRoom>> rooms)
+        : staff_{std::move(staff)}, rooms_{std::move(rooms)}
+    {
+    }
+};
