+++ modules/clinic/staff/doctor/i_doctor.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+
+class iDoctor
+{
+  public:
+    virtual ~iDoctor() = 0;
+};
