+++ modules/clinic/staff/doctor/i_serde_doctor.hpp
@@ -1,0 +1,1 @@
+#pragma once
