+++ modules/clinic/staff/doctor/i_visit_doctor.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+
+class iVisitDoctor : public iDoctor
+{
+  public:
+    virtual ~iVisitDoctor() = 0;
+};
