+++ modules/clinic/staff/i_staff.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+
+class iStaff
+{
+  public:
+    virtual ~iStaff() = 0;
+};
