+++ modules/clinic/staff/room/i_room.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+
+class iRoom
+{
+  public:
+    virtual ~iRoom() = 0;
+};
