+++ modules/common/enumerations/gender.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <cstdint>
+
+enum class Gender : std::uint8_t
+{
+    Male = 0,
+    Female = 1,
+    Unknown = 255
+};
