+++ modules/common/enumerations/treatment_state.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <cstdint>
+
+enum class TreatmentState : std::uint8_t
+{
+    Planned = 0,
+    Done = 1
+};
