+++ modules/common/person/address/address.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+struct Address
+{
+    std::string city_;
+    std::string country_;
+    std::string province_;
+    std::string street_;
+    std::string zip_code_;
+};
