+++ modules/common/person/name/name.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+struct Name
+{
+    std::string first_name_;
+    std::string second_name_;
+    std::string last_name_;
+};
