+++ modules/common/person/phone_number/phone_number.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+struct PhoneNumber
+{
+    std::string phone_code_; // +1, +48 (+ == 00)
+    std::string cellphone_number_;
+    std::string home_number;
+};
