+++ modules/patients/patient/i_serde_patient.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+// find better name for the interface
+class iSerdePatient : public virtual iPatient
+{
+  public:
+    virtual void SerializeData() = 0;
+    virtual ~iSerdePatient() = 0;
+};
